requirements.txt
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
imageio==2.9.0
importlib-metadata==3.1.0
imutils==0.5.3
ipykernel==4.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mock==4.0.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
notebook==5.2.2
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
Pillow==8.0.1
portpicker==1.2.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
tensorboard==1.13.1
tensorboard-plugin-wit==1.7.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
urllib3==1.26.2
waitress==1.4.4
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
